% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rotate_trends.R
\name{rotate_trends}
\alias{rotate_trends}
\title{rotate_trends is the primary function which calls pre-written stan scripts for time series data.}
\usage{
rotate_trends(fitted_model)
}
\arguments{
\item{fitted_model}{A fitted stanfit object}
}
\value{
a list object, with the rotated trends for each MCMC chain, along with the mean and 95% CIs
}
\description{
rotate_trends is the primary function which calls pre-written stan scripts for time series data.
}

